+++ emulator/slothpu/interface/_slothpu_interface.py
@@ -6,7 +6,7 @@ from ._backplane_widget import BackPlaneWidget
 from ._memory_column import MemoryColumn
 from ._pipeline_stage_widget import PipelineStageWidget
 from ._control_column import ControlColumn
-from._register_file_widget import RegisterFileWidget
+from ._register_column import RegisterColumn
 
 
 def top_handler(key):
@@ -14,21 +14,6 @@ def top_handler(key):
         raise urwid.ExitMainLoop()
 
 
-class RegisterColumn(urwid.WidgetWrap):
-    def __init__(self, target: SlothPU):
-        self._registers = RegisterFileWidget(target.register_file)
-
-        # Have to use Filler or urwid gets unhappy
-        register_pile = urwid.Pile(
-            [
-                urwid.Filler(self._registers, valign=urwid.TOP),
-            ]
-        )
-
-        super(RegisterColumn, self).__init__(register_pile)
-
-    def update(self):
-        self._registers.update()
 
 
 class SlothPU_Interface:
